Create README.md
This is the Task3 of Level-1 of the Oasis Infobyte Internship. Here, in this task, I built a temperature converter website where the user will have to input a temerature value that should be in Celsius, Fahrenheit or Kelvin and then press the converted button and the converted temperature will then be displayed with the correct unit.